+++ ec2_metadata.py
@@ -83,5 +83,10 @@ class EC2Metadata(object):
     def reservation_id(self):
         return requests.get(METADATA_URL + 'reservation-id').text
 
+    @cached_property
+    def security_groups(self):
+        # since this can be one or more items, always return as an array.
+        return requests.get(METADATA_URL + 'security-groups').text.splitlines()
+
 
 ec2_metadata = EC2Metadata()