+++ test_ec2_metadata.py
@@ -138,3 +138,21 @@ def test_region(resps):
 def test_reservation_id(resps):
     add_response(resps, 'reservation-id', 'r-12345678901234567')
     assert ec2_metadata.reservation_id == 'r-12345678901234567'
+
+def test_security_groups(resps):
+    # most common case: a single SG
+    add_response(resps, 'security-groups', 'security-group-one')
+    assert ec2_metadata.security_groups == ['security-group-one']
+    # another common case: multiple SGs
+    ec2_metadata.clear_all()
+    add_response(resps, 'security-groups', "security-group-one\nsecurity-group-2")
+    assert ec2_metadata.security_groups == ['security-group-one', 'security-group-2']
+    # check '' too. Can't create an instance without a SG but we should safely handle it,
+    # perhaps it's possible in OpenStack.
+    ec2_metadata.clear_all()
+    add_response(resps, 'security-groups', '')
+    assert ec2_metadata.security_groups == []
+    # finally, check None
+    ec2_metadata.clear_all()
+    add_response(resps, 'security-groups', None)
+    assert ec2_metadata.security_groups == []