+++ HISTORY.rst
@@ -7,6 +7,7 @@ Pending Release
 ---------------
 
 .. Insert new release notes below this line
+* Add security groups attribute.
 
 1.0.0 (2017-06-16)
 ------------------